+++ lib/interfaces/link.ts
@@ -1,8 +1,9 @@
+import {Observable} from 'rxjs/Observable';
 import {Collection} from './collection';
 import {LinkJSON} from './json';
 
 export interface LinkAPI {
-    follow(): Promise<Collection>;
+    follow(): Observable<Collection>;
 }
 
 export interface Link extends LinkJSON, LinkAPI {}